+++ src/OpenLoco/include/OpenLoco/Graphics/SoftwareDrawingContext.h
@@ -28,13 +28,13 @@ namespace OpenLoco::Gfx
         const RenderTarget& currentRenderTarget() const override;
         void clear(uint32_t fill) override;
         void clearSingle(uint8_t paletteId) override;
-        void fillRect(int16_t left, int16_t top, int16_t right, int16_t bottom, uint8_t colour, RectFlags flags) override;
-        void drawRect(int16_t x, int16_t y, uint16_t dx, uint16_t dy, uint8_t colour, RectFlags flags) override;
-        void fillRectInset(int16_t left, int16_t top, int16_t right, int16_t bottom, AdvancedColour colour, RectInsetFlags flags) override;
-        void drawRectInset(int16_t x, int16_t y, uint16_t dx, uint16_t dy, AdvancedColour colour, RectInsetFlags flags) override;
+        void fillRect(int32_t left, int32_t top, int32_t right, int32_t bottom, uint8_t colour, RectFlags flags) override;
+        void drawRect(int32_t x, int32_t y, int32_t dx, int32_t dy, uint8_t colour, RectFlags flags) override;
+        void fillRectInset(int32_t left, int32_t top, int32_t right, int32_t bottom, AdvancedColour colour, RectInsetFlags flags) override;
+        void drawRectInset(int32_t x, int32_t y, int32_t dx, int32_t dy, AdvancedColour colour, RectInsetFlags flags) override;
         void drawLine(const Ui::Point& a, const Ui::Point& b, PaletteIndex_t colour) override;
         void drawCircle(const Ui::Point& centre, int32_t radius, int32_t lineWidth, PaletteIndex_t colour) override;
-        void drawImage(int16_t x, int16_t y, uint32_t image) override;
+        void drawImage(int32_t x, int32_t y, uint32_t image) override;
         void drawImage(const Ui::Point& pos, const ImageId& image) override;
         void drawImageMasked(const Ui::Point& pos, const ImageId& image, const ImageId& maskImage) override;
         void drawImageSolid(const Ui::Point& pos, const ImageId& image, PaletteIndex_t paletteIndex) override;
